Switch to new cred method
