+++ type.go
@@ -55,6 +55,8 @@ type UserCredentialStore interface {
 type Settings struct {
 	// APIKey is the user's session key.
 	APIKey string
+	// Credential holds the user's credential data.
+	Credential *Credential
 }
 
 // Credential is a struct that holds credential information.
